feat: add 16-bit ADD HL,n instruction with unit tests

+++ test/alu_8bit_sub_test.c
@@ -42,7 +42,7 @@ Test(sub_a_n, sub_a_a_no_borrow, .init = cpu_mmu_setup, .fini = cpu_teardown) {
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, (address + 1)));
+    cr_expect(eq(u16, cpu.PC, (address + 1)));
 }
 
 Test(sub_a_n, sub_a_b_only_half_borrow, .init = cpu_mmu_setup, .fini = cpu_teardown) {
@@ -70,7 +70,7 @@ Test(sub_a_n, sub_a_b_only_half_borrow, .init = cpu_mmu_setup, .fini = cpu_teard
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, address + 1));
+    cr_expect(eq(u16, cpu.PC, address + 1));
 }
 
 Test(sub_a_n, sub_a_c_borrow, .init = cpu_mmu_setup, .fini = cpu_teardown) {
@@ -98,7 +98,7 @@ Test(sub_a_n, sub_a_c_borrow, .init = cpu_mmu_setup, .fini = cpu_teardown) {
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, address + 1));
+    cr_expect(eq(u16, cpu.PC, address + 1));
 }
 
 Test(sub_a_n, sub_a_d_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
@@ -114,7 +114,7 @@ Test(sub_a_n, sub_a_d_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, address + 1));
+    cr_expect(eq(u16, cpu.PC, address + 1));
 }
 
 Test(sub_a_n, sub_a_e_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
@@ -130,7 +130,7 @@ Test(sub_a_n, sub_a_e_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, address + 1));
+    cr_expect(eq(u16, cpu.PC, address + 1));
 }
 
 Test(sub_a_n, sub_a_h_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
@@ -146,7 +146,7 @@ Test(sub_a_n, sub_a_h_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, address + 1));
+    cr_expect(eq(u16, cpu.PC, address + 1));
 }
 
 Test(sub_a_n, sub_a_l_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
@@ -162,7 +162,7 @@ Test(sub_a_n, sub_a_l_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, address + 1));
+    cr_expect(eq(u16, cpu.PC, address + 1));
 }
 
 Test(sub_a_n, sub_a_hl_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
@@ -189,7 +189,7 @@ Test(sub_a_n, sub_a_hl_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, address + 1));
+    cr_expect(eq(u16, cpu.PC, address + 1));
 }
 
 Test(sub_a_n, sub_a_d8_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
@@ -214,5 +214,5 @@ Test(sub_a_n, sub_a_d8_res_only, .init = cpu_mmu_setup, .fini = cpu_teardown) {
     cr_expect(eq(u8, actual, expected));
 
     // check if PC is updated correctly
-    cr_expect(eq(u8, cpu.PC, address + 2));
+    cr_expect(eq(u16, cpu.PC, address + 2));
 }
